+++ MoMoney.Core/Models/DetailedStock.cs
@@ -1,9 +1,10 @@
 ﻿
 namespace MoMoney.Core.Models;
 
-public class DetailedStock : Stock
+public partial class DetailedStock : Stock
 {
     public decimal MarketValue => Symbol.EndsWith(":TSE") ? MarketPrice * Quantity : MarketPrice * Quantity * 1.35m;
+    public decimal BookValue => Quantity * Cost;
     public decimal Change => MarketValue - BookValue;
     public decimal ChangePercent => (MarketValue / BookValue) - 1;
 
@@ -13,6 +14,5 @@ public class DetailedStock : Stock
         Quantity = stock.Quantity;
         Cost = stock.Cost;
         MarketPrice = stock.MarketPrice;
-        BookValue = stock.BookValue;
     }
 }