+++ MoMoney.Core/Models/Stock.cs
@@ -7,22 +7,20 @@ public partial class Stock : ObservableObject
 {
     [PrimaryKey]
     public string Symbol { get; set; }
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public decimal Cost { get; set; }
 
     [ObservableProperty]
     public decimal marketPrice;
-    public decimal BookValue { get; set; }
 
     public Stock() { }
 
-    public Stock(string symbol, int quantity, decimal cost, decimal marketPrice, decimal bookValue)
+    public Stock(string symbol, decimal quantity, decimal cost, decimal marketPrice)
     {
         Symbol = symbol;
         Quantity = quantity;
         Cost = cost;
         MarketPrice = marketPrice;
-        BookValue = bookValue;
     }
 
     public Stock (Stock stock)
@@ -31,6 +29,5 @@ public partial class Stock : ObservableObject
         Quantity = stock.Quantity;
         Cost = stock.Cost;
         MarketPrice = stock.MarketPrice;
-        BookValue = stock.BookValue;
     }
 }