+++ MoMoney/Views/Settings/LoggingPage.xaml.cs
@@ -10,7 +10,7 @@ public partial class LoggingPage : ContentPage
 		InitializeComponent();
         BindingContext = vm;
 		vm.listview = listview;
-		Loaded += vm.Init;
+		Loaded += async (s, e) => await vm.Init();
 
 		PckLevels.SelectedValueChanged += (s, e) =>
 		{